Modularize auth handlers
Feat: Delete cookie if unavailable
